requirements.txt
streamlit
numpy
pillow
plotly
